requirements.txt
APScheduler==3.10.4
firebase-admin==6.5.0
firebase-auth==3.0.28
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
google-api-core==2.21.0
google-api-python-client==2.42.0
google-auth==2.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
Werkzeug==2.3.7
gunicorn==20.1.0
celery==5.3.0
python-dotenv==1.0.0